requirements.txt
asgiref==3.8.1
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
uuid==1.30
